move util func
